refactor(assistant): split Workers AI gateway handler test
- move AI Gateway routing coverage into a Workers AI provider-local test
- leave handler.test focused on tool streaming and browser UI message handling
- reduce handler.test size from 216 to 167 lines while keeping Fallow clean
